LayoutTests/accessibility/mac/session-id.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that we can receive the page ID from the web area element.");

    if (window.accessibilityController) {

          var webArea = accessibilityController.rootElement.childAtIndex(0);
          shouldBe("webArea.role", "'AXRole: AXWebArea'");
          shouldBeTrue("webArea.numberAttributeValue('AXWebSessionID') >= 1");
    }

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
