LayoutTests/accessibility/mac/session-id-expected.txt

This tests that we can receive the page ID from the web area element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS webArea.role is 'AXRole: AXWebArea'
PASS webArea.numberAttributeValue('AXWebSessionID') >= 1 is true
PASS successfullyParsed is true

TEST COMPLETE
